refactor(elbv2): migrate listener and rule code to _elbv2 submodules
Extract listener and rule management functionality from elbv2.py into
dedicated submodules for better code organization:

- Create _elbv2/actions.py (176 lines) - action processing utilities
- Create _elbv2/listeners.py (251 lines) - listener management
- Create _elbv2/rules.py (401 lines) - rule management
- Reduce elbv2.py from 1,418 to 753 lines
- Maintain backward compatibility via re-exports

Reorganize tests to mirror source structure:
- test_prune.py → _elbv2/test_actions.py
- Extract listener tests → _elbv2/test_listeners.py
- Merge rule tests → _elbv2/test_rules.py

All 216 elbv2 unit tests pass with the new structure.
